move dialogs to global level
