+++ src/layouts/default/PlayerOSD/PlayerTrackDetails.vue
@@ -127,7 +127,7 @@ import api from '@/plugins/api';
 import { MediaType, MediaItemType, ItemMapping } from '@/plugins/api/interfaces';
 import { store } from '@/plugins/store';
 import MediaItemThumb from '@/components/MediaItemThumb.vue';
-import { getArtistsString } from '@/utils';
+import { getArtistsString } from '@/helpers/utils';
 import { useRouter } from 'vue-router';
 import PlayerFullscreen from './PlayerFullscreen.vue';
 import { iconFallback } from '@/components/QualityDetailsBtn.vue';
