+++ src/views/ArtistDetails.vue
@@ -84,7 +84,7 @@ import ProviderIcon from '@/components/ProviderIcon.vue';
 import { api } from '../plugins/api';
 import ListItem from '../components/mods/ListItem.vue';
 import Container from '../components/mods/Container.vue';
-import { getStreamingProviderMappings } from '@/utils';
+import { getStreamingProviderMappings } from '@/helpers/utils';
 
 export interface Props {
   itemId: string;
