+++ src/views/PlayerQueue.vue
@@ -184,7 +184,7 @@ import type { QueueItem, EventMessage, MediaItemType } from '../plugins/api/inte
 import { EventType, MediaType } from '../plugins/api/interfaces';
 import { api } from '../plugins/api';
 import { store } from '../plugins/store';
-import { truncateString } from '../utils';
+import { truncateString } from '@/helpers/utils';
 import { useI18n } from 'vue-i18n';
 import { useRouter } from 'vue-router';
 import ListviewItem from '@/components/ListviewItem.vue';
@@ -354,4 +354,3 @@ watch(
 );
 </script>
 
-<style scoped></style>
